aoc_2025_09

Module visibility

Visibility Module

This module provides functionality to determine visibility bounds from a given point within a polygon defined by a set of indexed coordinates. It includes functions to check if a line blocks visibility, find visibility bounds, and build visibility bounds for a list of indexed coordinates.

## Warning

This module does not work:

..............
.......#XXX#..
.......XXXXX..
..AXXXXBXXXC..
..XXXXXXXXXX..
..#XXXXXX#XX..
.........XXX..
.........#X#..
..............

A is at (2,3), B is at (7,3), C is at (11,3).

Since visibility lines has no concept of “inside” or “outside” the polygon, it consider B to be blocking A’s visibility to the right, while this is only true if the polygon is flipped:

XXXXXXXXXXXXXX
XXXXXXX#XXX#XX
XXXXXXXX...XXX
XXAXXXXB...CXX
XXX........XXX
XX#XXXXXX#.XXX
XXXXXXXXXX.XXX
XXXXXXXXX#X#XX
XXXXXXXXXXXXXX

In other words, the visibility from A to C is dependant on whether the polygon is considered to be the “X” area or the “.” area; but this module has no concept of that.

Some of this problem was mitigated by using two visibility bounds (absolute and corner): if we hit a corner instead of a straight edge, we treat that as a weaker form of blocking; and behave the opposite way when combining the two bounds:

for absolute bounds, we take the closest blocking edge we find; but

for corner bounds, we take the furthest blocking edge we find.

This however is not sufficient to solve the problem in all cases:

..............
.AXXXXXXXXXX#.
.BXXXXXX#XXXX.
........XXXXX.
.CXXXXXX#XXXX.
.DXXXXXXXXXX#.
..............

In the above example, A only encounters corners when looking downwards. Using our logic above, it sees B, C and D, and prefers D as the bottom visibility bound. However, in reality, there is a gap after B that C and D should not be visible.

Without knowing which side of the polygon is “inside” or “outside”, this cannot be correctly resolved.

In actual fact, our input happens to not trigger this limitation, so the answer is still correct, but this is absolute garbage code that only works by coincidence.

Enums

Direction

Functions

blocks_visibility 🔒

Determines if a line defined by two IndexedCoords blocks visibility from a given point.

build_visibility_bounds_for_indexed_coords

Builds visibility bounds for a list of IndexedCoords based on a polygon.

find_visibility_bounds

Finds the visibility bounds from a given coordinate within a polygon defined by the IndexedCoords.